+++ src/YarpUtilities/src/Helper.cpp
@@ -54,3 +54,60 @@ template <> bool YarpUtilities::convertValue<bool>(const yarp::os::Value& value)
 {
     return value.asBool();
 }
+
+template <>
+bool YarpUtilities::getVectorFromSearchable<std::vector<bool>>(const yarp::os::Searchable& config,
+                                                               const std::string& key,
+                                                               std::vector<bool>& vector)
+{
+    yarp::os::Value* value;
+    if (!config.check(key, value))
+    {
+        std::cerr << "[BipedalLocomotionControllers::YarpUtilities::getVectorFromSearchable] "
+                     "Missing field "
+                  << key << std::endl;
+        return false;
+    }
+
+    if (value->isNull())
+    {
+        std::cerr << "[BipedalLocomotionControllers::YarpUtilities::getVectorFromSearchable] Empty "
+                     "input value named "
+                  << key << std::endl;
+        return false;
+    }
+
+    if (!value->isList())
+    {
+        std::cerr << "[BipedalLocomotionControllers::YarpUtilities::getVectorFromSearchable] The "
+                     "value named "
+                  << key << "is not associated to a list." << std::endl;
+        return false;
+    }
+
+    yarp::os::Bottle* inputPtr = value->asList();
+    if (inputPtr == nullptr)
+    {
+        std::cerr << "[BipedalLocomotionControllers::YarpUtilities::getVectorFromSearchable] The "
+                     "list associated to the value named "
+                  << key << " is empty." << std::endl;
+        return false;
+    }
+
+    // resize the vector
+    vector.resize(inputPtr->size());
+
+    for (int i = 0; i < inputPtr->size(); i++)
+    {
+        if (!(inputPtr->get(i).isBool()) && !(inputPtr->get(i).isInt()))
+        {
+            std::cerr << "[BipedalLocomotionControllers::YarpUtilities::getVectorFromSearchable] "
+                         "The element of the list associated to the value named "
+                      << key << " is not a boolean ." << std::endl;
+            return false;
+        }
+
+        vector[i] = YarpUtilities::convertValue<bool>(inputPtr->get(i));
+    }
+    return true;
+}