+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project are documented in this file.
 - The `cmake/InstallBasicPackageFiles.cmake` file
 - Implement the first version of the `BipedalLocomotionControllers` interface
 - Implement the  first version of the `YarpUtilities` library
+- Implement `ParametersHandler` library (https://github.com/dic-iit/bipedal-locomotion-controllers/pull/13)
 - Implement the first version of the `OptimalControlUtilities` library
 
 [Unreleased]: https://github.com/dic-iit/bipedal-locomotion-controllers/